# packages/backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/consultants": {
      "post": {
        "operationId": "ConsultantsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConsultantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Consultant/User created successfully."
          },
          "403": {
            "description": "Forbidden. Only admins can create users here."
          },
          "409": {
            "description": "Conflict. User with this email already exists."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Consultants Management"
        ]
      },
      "get": {
        "operationId": "ConsultantsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Field to sort by",
            "schema": {
              "default": "createdAt",
              "example": "name",
              "type": "string",
              "enum": [
                "name",
                "email",
                "lastActive",
                "createdAt",
                "status"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order",
            "schema": {
              "default": "asc",
              "example": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc",
                "ASC",
                "DESC"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search by name or email",
            "schema": {
              "example": "Doe",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by status",
            "schema": {
              "example": "active",
              "type": "string",
              "enum": [
                "active",
                "inactive"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved list of consultants."
          },
          "403": {
            "description": "Forbidden. Only admins can list consultants."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Consultants Management"
        ]
      }
    },
    "/api/consultants/{id}": {
      "get": {
        "operationId": "ConsultantsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved consultant details."
          },
          "403": {
            "description": "Forbidden. Only admins can view consultant details."
          },
          "404": {
            "description": "Consultant not found or is not a consultant."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Consultants Management"
        ]
      },
      "put": {
        "operationId": "ConsultantsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateConsultantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Consultant/User updated successfully."
          },
          "403": {
            "description": "Forbidden. Only admins can update users here."
          },
          "404": {
            "description": "User not found or is not a manageable role."
          },
          "409": {
            "description": "Conflict. User with this email already exists."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Consultants Management"
        ]
      }
    },
    "/api/consultants/{id}/reset-password": {
      "post": {
        "operationId": "ConsultantsController_resetPassword",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetConsultantPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password for the user has been successfully reset."
          },
          "403": {
            "description": "Forbidden. Only admins can reset passwords."
          },
          "404": {
            "description": "User not found or not a role whose password can be reset."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Consultants Management"
        ]
      }
    },
    "/api/consultants/{id}/inventory": {
      "get": {
        "operationId": "ConsultantsController_getConsultantInventoryVisibility",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Consultant ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of inventory items with their visibility status for the consultant.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConsultantInventoryVisibilityDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Consultant not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get inventory visibility for a specific consultant (Admin only)",
        "tags": [
          "Consultants Management"
        ]
      },
      "put": {
        "operationId": "ConsultantsController_updateConsultantInventoryVisibility",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Consultant ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateConsultantInventoryVisibilityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Consultant inventory visibility updated successfully."
          },
          "400": {
            "description": "Invalid input data."
          },
          "404": {
            "description": "Consultant or some inventory items not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update inventory visibility for a specific consultant (Admin only)",
        "tags": [
          "Consultants Management"
        ]
      }
    },
    "/api/consultants/{id}/appointment-types": {
      "get": {
        "operationId": "ConsultantsController_getConsultantAppointmentTypesVisibility",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Consultant ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of appointment types with their visibility status for the consultant.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConsultantAppointmentTypeVisibilityDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Consultant not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get appointment types visibility for a specific consultant (Admin only)",
        "tags": [
          "Consultants Management"
        ]
      },
      "put": {
        "operationId": "ConsultantsController_updateConsultantAppointmentTypesVisibility",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Consultant ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateConsultantAppointmentTypesVisibilityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Consultant appointment types visibility updated successfully."
          },
          "400": {
            "description": "Invalid input data."
          },
          "404": {
            "description": "Consultant or some appointment types not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update appointment types visibility for a specific consultant (Admin only)",
        "tags": [
          "Consultants Management"
        ]
      }
    },
    "/api/consultants/{id}/stats": {
      "get": {
        "operationId": "ConsultantsController_getConsultantStats",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Consultant ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved consultant statistics.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsultantStatsDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden. Only admins can access this resource."
          },
          "404": {
            "description": "Consultant not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get statistics for a specific consultant (Admin only)",
        "tags": [
          "Consultants Management"
        ]
      }
    },
...
  },
  "info": {
    "title": "Supplement Management API",
    "description": "API documentation for the Supplement Management System",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateConsultantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe",
            "description": "Full name of the consultant"
          },
          "email": {
            "type": "string",
            "example": "john.consultant@example.com",
            "description": "Email address (must be unique)"
          },
          "password": {
            "type": "string",
            "example": "Str0ngP@sswOrd",
            "description": "Password (at least 8 characters)"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "consultant"
            ],
            "description": "Role of the user. For this endpoint, typically 'consultant' or 'admin'.",
            "example": "consultant"
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "role"
        ]
      },
      "UpdateConsultantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Johnathan Doe",
            "description": "Full name of the consultant"
          },
          "email": {
            "type": "string",
            "example": "john.doe.new@example.com",
            "description": "Email address (must be unique)"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "consultant"
            ],
            "description": "Role of the user",
            "example": "consultant"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive"
            ],
            "description": "Status of the consultant account",
            "example": "active"
          }
        }
      },
      "ResetConsultantPasswordDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "example": "NewStr0ngP@sswOrd",
            "description": "New password (at least 8 characters)"
          }
        },
        "required": [
          "password"
        ]
      },
      "ConsultantInventoryVisibilityDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "ID of the inventory item"
          },
          "name": {
            "type": "string",
            "example": "Vitamin C 500mg",
            "description": "Name of the inventory item"
          },
          "visible": {
            "type": "boolean",
            "example": true,
            "description": "Whether the item is visible to the consultant"
          }
        },
        "required": [
          "id",
          "name",
          "visible"
        ]
      },
      "UpdateConsultantInventoryVisibilityDto": {
        "type": "object",
        "properties": {
          "inventoryItemIds": {
            "description": "Array of inventory item IDs to be made visible to the consultant. Replaces existing specific visibilities.",
            "example": [
              1,
              2,
              3
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "inventoryItemIds"
        ]
      },
      "ConsultantAppointmentTypeVisibilityDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "The ID of the appointment type"
          },
          "name": {
            "type": "string",
            "example": "Initial Consultation",
            "description": "The name of the appointment type"
          },
          "visible": {
            "type": "boolean",
            "example": true,
            "description": "Whether the appointment type is visible to the consultant"
          }
        },
        "required": [
          "id",
          "name",
          "visible"
        ]
      },
      "UpdateConsultantAppointmentTypesVisibilityDto": {
        "type": "object",
        "properties": {
          "appointmentTypeIds": {
            "description": "An array of appointment type IDs that should be visible to the consultant.",
            "example": [
              1,
              2,
              3
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "appointmentTypeIds"
        ]
      },
      "AppointmentByTypeStatsDto": {
        "type": "object",
        "properties": {
          "typeId": {
            "type": "number",
            "example": 1,
            "description": "ID of the appointment type"
          },
          "typeName": {
            "type": "string",
            "example": "Initial Consultation",
            "description": "Name of the appointment type"
          },
          "count": {
            "type": "number",
            "example": 10,
            "description": "Number of appointments of this type"
          }
        },
        "required": [
          "typeId",
          "typeName",
          "count"
        ]
      },
      "RecentAppointmentStatsDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 101,
            "description": "ID of the appointment"
          },
          "patientName": {
            "type": "string",
            "example": "John Doe",
            "description": "Name of the patient"
          },
          "date": {
            "type": "string",
            "example": "2024-07-30T10:00:00.000Z",
            "description": "Date and time of the appointment"
          },
          "typeName": {
            "type": "string",
            "example": "Follow-up",
            "description": "Name of the appointment type"
          }
        },
        "required": [
          "id",
          "patientName",
          "date",
          "typeName"
        ]
      },
      "ConsultantStatsDto": {
        "type": "object",
        "properties": {
          "totalPatients": {
            "type": "number",
            "example": 50,
            "description": "Total number of patients assigned to the consultant"
          },
          "totalAppointments": {
            "type": "number",
            "example": 120,
            "description": "Total number of appointments conducted by the consultant"
          },
          "totalRevenue": {
            "type": "number",
            "example": 15000.5,
            "description": "Total revenue generated from the consultant's appointments"
          },
          "appointmentsByType": {
            "description": "Breakdown of appointments by type",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AppointmentByTypeStatsDto"
            }
          },
          "recentAppointments": {
            "description": "List of recent appointments",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecentAppointmentStatsDto"
            }
          }
        },
        "required": [
          "totalPatients",
          "totalAppointments",
          "totalRevenue",
          "appointmentsByType",
          "recentAppointments"
        ]
      }
    }
  }
}